+++ src/extension/formatter/formatter.ts
@@ -2,7 +2,7 @@ import * as fs from "fs";
 import * as path from "path";
 import * as vs from "vscode";
 
-import { formatterPath as cformatterPath } from "../../shared/constants";
+import { formatterServerPath as cformatterServerPath } from "../../shared/constants";
 import { Logger } from "../../shared/interfaces";
 import { extensionPath, extensionVersion } from "../../shared/vscode/extension_utils";
 import { isRunningLocally } from "../../shared/vscode/utils";
@@ -10,31 +10,31 @@ import { config } from "../config";
 
 
 export function getFormatterArgs(logger: Logger) {
-	/*const formatterPath = config.formatterPath || (
+	/*const formatterServerPath = config.formatterServerPath || (
 		dartCapabilities.supportsLanguageServerCommand
 			? "language-server"
 			: path.join(sdks.dart, formatterSnapshotPath)
 	);*/
 	//TODO: for now use binaries for each platform.
 	// see if we can use dill and get a dart sdk path.
-	const formatterPath = config.formatterPath || path.join(extensionPath, cformatterPath);
+	const formatterServerPath = config.formatterServerPath || path.join(extensionPath, cformatterServerPath);
 
 	// If the ssh host is set, then we are running the formatter on a remote machine, that same formatter
 	// might not exist on the local machine.
-	if (!config.formatterSshHost && !fs.existsSync(formatterPath)) {
-		const msg = "Could not find a Dart Formatter Server at " + formatterPath;
+	if (!config.formatterSshHost && !fs.existsSync(formatterServerPath)) {
+		const msg = "Could not find a Dart Formatter Server at " + formatterServerPath;
 		vs.window.showErrorMessage(msg);
 		logger.error(msg);
 		throw new Error(msg);
 	}
 
-	return buildFormatterArgs(formatterPath);
+	return buildFormatterArgs(formatterServerPath);
 }
 
-function buildFormatterArgs(formatterPath: string) {
+function buildFormatterArgs(formatterServerPath: string) {
 	let formatterArgs = [];
 
-	formatterArgs.push(formatterPath);
+	formatterArgs.push(formatterServerPath);
 
 	// formatter in server mode,
 	formatterArgs.push("listen");