+++ src/shared/constants.ts
@@ -19,7 +19,7 @@ export const executableNames = {
 };
 export const getExecutableName = (cmd: string) => (executableNames as { [key: string]: string | undefined })[cmd] ?? cmd;
 export const dartVMPath = "bin/" + executableNames.dart;
-export const formatterPath = "bin/" + executableNames.dartFormatter;
+export const formatterServerPath = "bin/" + executableNames.dartFormatter;
 
 export const stopLoggingAction = "Stop Logging";
 export const showLogAction = "Show Log";